updated eclipse file w/server
i have now got the server working on http//:localhost/8005. The server code is currently commented out to test the CRUD menu. For some reason the database won't show up in the path folder on eclipse so the menu currently won't retrieve any data until it is fixed. As of now the CRUD menu has the ability to: retrieve all Pokemon, search for a Pokemon by ID, and search for a Pokemon by name.